requirements.txt
coverage==4.3.4
django==1.10.6
django-prometheus==1.0.9
django-bootstrap3==8.2.1
sw-django-utils==0.0.35
sw-python-utils==0.0.15
djangorestframework==3.6.2
gunicorn==19.6.0
mock==2.0.0
factory-boy==2.8.1
croniter==0.3.16
django-tz-detect==0.2.8
notify-client==0.0.3